Rename useGetCleanBranches.tsx to branches.tsx
Because it contains exported functions besides the useGetCleanBranches hook and other files use these exported functions as well.
